feat(qa): 改进测试用例版本显示和评审记录处理
- 在侧边栏面板中显示具体版本号而非"最新"
- 修改测试用例版本字段为浮点数以支持更细粒度的版本控制
- 优化评审记录更新逻辑，避免重复记录相同评审结果
- 调整评审界面样式和布局
- 添加更新时间字段到评审记录

+++ apps/api/plane/app/views/qa/case_version.py
@@ -39,8 +39,8 @@ class CaseVersionCompareAPIView(BaseAPIView):
             )
 
         try:
-            from_version_int = -1 if str(from_version).lower() == "current" else int(from_version)
-            to_version_int = -1 if str(to_version).lower() == "current" else int(to_version)
+            from_version_int = -1 if str(from_version).lower() == "current" else float(from_version)
+            to_version_int = -1 if str(to_version).lower() == "current" else float(to_version)
         except ValueError:
             return Response(
                 {"error": "from_version and to_version must be integers"},
@@ -124,5 +124,3 @@ class CaseVersionCompareAPIView(BaseAPIView):
             }
         )
         return Response(serializer.data, status=status.HTTP_200_OK)
-
-
